防止重复文件通过网页上传，生成多条记录

+++ protected/plugins/miniStore/biz/PluginMiniStoreBiz.php
@@ -16,17 +16,21 @@ class PluginMiniStoreBiz extends MiniBiz{
      * @param int $nodeId 迷你存储节点值 
      */
     public function report($path,$signature,$size,$nodeId){
-        //创建version/versionMeta数据
-        $pathParts = pathinfo($path);
-        $type = CUtils::mime_content_type($pathParts["filename"]);
-        $version = MiniVersion::getInstance()->create($signature, $size, $type);
-        MiniVersionMeta::getInstance()->create($version["id"],"store_id",$nodeId);
-        //更新迷你存储节点状态，把新上传的文件数+1
-        PluginMiniStoreNode::getInstance()->newUploadFile($nodeId);
-        //文档转换
-        do_action('file_upload_after', $signature);
-        //清理垃圾数据
-        PluginMiniBreakFile::getInstance()->deleteBySignature($signature);
+        //防止重复文件通过网页上传，生成多条记录
+        $version = MiniVersion::getInstance()->getBySignature($signature);
+        if(empty($version)){
+            //创建version/versionMeta数据
+            $pathParts = pathinfo($path);
+            $type = CUtils::mime_content_type($pathParts["filename"]);
+            $version = MiniVersion::getInstance()->create($signature, $size, $type);
+            MiniVersionMeta::getInstance()->create($version["id"],"store_id",$nodeId);
+            //更新迷你存储节点状态，把新上传的文件数+1
+            PluginMiniStoreNode::getInstance()->newUploadFile($nodeId);
+            //文档转换
+            do_action('file_upload_after', $signature);
+            //清理垃圾数据
+            PluginMiniBreakFile::getInstance()->deleteBySignature($signature);
+        }
         //执行文件秒传逻辑
         $filesController = new MFileSecondsController();
         $filesController->invoke();
